print-events/requirements.txt
PyJWT[cryptography]
cryptography
awslambdaric
boto3
